Adding a sample runner class.
In order to make the runner more testable, we would like to
encapsulate its logic into a class. This allows us to run tests
against its functionality.

The driving factor behind this change currently is the display of an
error message when a check fails.

The organization of the ConsoleLog check requires an error message
to be output in the case of a failure.

This change was made in the ConsoleLog check in the first place to
make testing more pleasant. The old version spit out an error message
immediately upon failure, which is fine at runtime. However, when
testing error cases, we don't want to clutter the test output with
user error messages.