Only checking lines that the committer has added.
The debugger check was causing some issues with certain pieces of
code. e.g. a file contained code that looked like:

    def debugger
    end

If this was the case, the debugger check would always fail this file,
even if the committer didn't touch code anywhere near that method
definition. This change allows committers to freely modify files
containing the word "debugger" as long as they are not adding it
themselves.