Fixing Jslint checks.
The Jslint checks (or at least the "new" check) was always returning
a failure. The way it was built, it added errors to an array. When
the check was done, if the array was not empty the check failed.

The problem was that the check also returned an empty array,
resulting in errors that looked like this:

    errors.inspect # => [[]]

So, now we're flattening `errors` before we check if it's empty.

    [[]].empty? # => false
    [].empty? # => true