Skipping Jslint if therubyracer is not installed.
Previously, if therubyracer was not installed on the system, the
Jslint check would fail, and thus the entire commit would fail.

We would start out by attempting to require 'v8' and if it failed,
we returned an empty array. This caused the entire check to fail. See
3ddc227836ac44224e96d62e2d431a605d99e26f for more information.

Now, instead of erring out and failing, we print a warning message
and skip the Jslint checks if therubyracer is not installed.

The rest of the pre-commit checks run fine, the developer is able to
commit, and everyone is happier! (: