Being safer about displaying errors.
When JSHint encounters a javascript file with a lot of errors in it,
it just gives up after it's scanned some percentage of the file.

In this case, it starts returning nil's instead of error hashes. When
we receive nil's the error reporting code blows up.

This should stop that behavior.