A more portable shebang for the pre-commit template.
The shebang currently tries to use a flavor of ruby and it also
tries to require rubygems.

Some flavors of Linux do not split up arguments on the shebang line.
For example, given the following shebang:

    #!/usr/bin/env ruby -rubygems

On one system the shebang may be interpreted as passing the argument
`-rubygems` to `ruby`. On another system the shebang may be
interpreted as passing `ruby -rubygems` to the `env` command.

This commit moves the responsibility of requiring rubygems into the
code and out of the shebang.

cf: http://en.wikipedia.org/wiki/Shebang_(Unix)#Portability