Checking for the presence of the pre-commit gem.
Sometimes people have a pre-commit hook installed that requires the
pre-commit gem, but they do not have the pre-commit gem installed.

Typically this is caused by installing the pre-commit gem and hook,
then upgrading to a newer version of Ruby.

The pre-commit hook still remains, but the new version of Ruby does
not yet have the pre-commit gem installed. This check is a
convenience to help guide users to re-install the gem for the new
version of Ruby.