Integration tests do not depend on implementation.
(hopefully)

We were counting on an implementation detail of the pre commit hook
template to run integration tests on ci servers. The hook used to
require the pre-commit library using the `-r` Ruby option.
Unfortunately old versions of Ruby do not support passing the `-r`
option multiple times. When we changed the hook to only pass one `-r`
option the integration tests started failing.

In order to get around this, we're hooking into a constant called
`OPTIONS`. I'm not sure if this is actually any better, as we're
still counting on the fact that the `OPTIONS` constant is defined in
the pre commit hook template.

To help alleviate some pain in the future, we have a nicer test
failure message if the `OPTIONS` constant is missing.