`Cli#install` handles missing second argument.
When a second argument is missing ARGV[x] returns `nil`. If we pass
`nil` into a method, that is the value, not the `key = foo` default.
Here we check for the lack of a key, and use "default" if one is not
present.

If this gets much more complex, we should switch to a proper option
parser.