Js and Grep test work with Guard.
We have a pattern in the Guardfile to watch when files under lib are
saved and then run the corresponding test. The `js.rb` and `guard.rb`
files were located in directory structures that did not match their
corresponding test files.

This moves the test files, so Guard will properly pick up changes
when the source files are changed.