The binary check handles unicode source files.
The binary? query makes a best guess as to whether a file is a binary
file or not. The previous check was reading some bytes form the
beginning of the file. It would then check to see what percentage of
those bytes were common ASCII characters. If the file was made up of
70% common ASCII characters, we would "guess" that the file was not
a binary file.

One of the failings of this test is that it does not take unicode
characters into account. To fix that, we attempt to convert all
unicode characters to the ASCII character `?` and then perform the
same calculation to see if 70% of the file contains common characters.