Do not run pre-commit checks on images
We have a mechanism to determine if we should run our list of checks
on the files being committed. Generally these checks are pretty
simplistic. For example, does the file exist, or is it a directory?

Also, in order to avoid running checks on large binary files we run
a heuristic against every file currently being committed.
Unfortunately, this binary check is prone to errors and is now causing
performance issues.

In order to avoid this binary file test, we do some quick filtering
based on file extension. If the file is an image, we definitely do not
want to run checks on it. If the file is a Ruby source code file, we
definitely do want to run checks on it.

This will allow us to avoid running the binary test in most cases.