Add more binary and source file extensions
Opening files and examining their contents to determine if they are
binary files or source files is expensive. In order to avoid some of
this extra overhead, we maintain two separate lists of file
extensions.

If a file extension is in the binary list, we assume it is binary and
we do not run any checks on it. If the file extension is in the source
file list then we do run checks on it.