Ignore standard error when checking grep version
We are currently shelling out to the `grep` command for some
functionality of the pre-commit gem. In order to do this we need to
reason about the `grep` executable running on the system.

To introspect the command we check the output of the `grep --version`
command and decide what to do next. On Alpine Linux grep does not
report its version and instead outputs an error message. Here we
ignore the standard error stream when checking grep's version.